Add files via upload
run it accordingly streamlit_dashboard isn't complete yet. Try audit_tool it fill finds that weather the value is present or not in the text file and give the output of the file how many empty spaces are there and how many duplicates are there in the given text file.